Split various types out of Socket.cs.
This change moves SocketAsyncEventArgs, SocketAsyncOperation, and
SendPacketsElement into their own files.
