+++ src/System.Net.Sockets/src/System.Net.Sockets.csproj
@@ -37,7 +37,10 @@
     <Compile Include="System\Net\Sockets\NetworkStream.cs" />
     <Compile Include="System\Net\Sockets\ProtocolType.cs" />
     <Compile Include="System\Net\Sockets\SelectMode.cs" />
+    <Compile Include="System\Net\Sockets\SendPacketsElement.cs" />
     <Compile Include="System\Net\Sockets\Socket.cs" />
+    <Compile Include="System\Net\Sockets\SocketAsyncEventArgs.cs" />
+    <Compile Include="System\Net\Sockets\SocketAsyncOperation.cs" />
     <Compile Include="System\Net\Sockets\SocketFlags.cs" />
     <Compile Include="System\Net\Sockets\SocketInformation.cs" />
     <Compile Include="System\Net\Sockets\SocketInformationOptions.cs" />
